Fixed an issue with foldernames. read desc.
The previous way of creating folders was <seriesName>/<counter>_<amiiboName> this created an issue where, if an amiibo was added somewhere in the middle of the api/database, the numbers would shit and multiple folders with the same amiibo but a different id exists.

Now i am using the amiibo id as suffix (which hopefully is unique). This way every amiibo will always have the same folder name, no matter the position in the database.